refactor: move modal component to core
