+++ src/app/core/modal/modal.component.spec.ts
@@ -3,7 +3,7 @@ import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { TranslateService, TranslateModule } from '@ngx-translate/core';
 import { ModalComponent } from './modal.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { MockComponent } from '../../core/mocks/mock-component';
+import { MockComponent } from '../mocks/mock-component';
 
 describe('ModalComponent', () => {
   let component: ModalComponent;
