+++ src/app/shared/shared.module.ts
@@ -11,7 +11,6 @@ import { DropDownComponent } from './dropdown/dropdown.component';
 import { ReactiveFormsModule } from '@angular/forms';
 import { SearchComponent } from './search/search.component';
 import { AngularSvgIconModule } from 'angular-svg-icon';
-import { ModalComponent } from './modal/modal.component';
 
 
 @NgModule({
@@ -22,7 +21,6 @@ import { ModalComponent } from './modal/modal.component';
     BadgeComponent,
     DropDownComponent,
     SearchComponent,
-    ModalComponent
   ],
   imports: [
     CommonModule,
